Move identityset into api package
Makes sense since this is an API library tool.
